+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - `org`: `Identity` now has `gln` as a possible Key.
 - `eu-en16931-v2017`: `Identity` normalization adds iso scheme codes extension for certain keys.
 - `ar`: Argentine regime
+- `it-sdi-v1`: Add Italian phone number validation and normalization
 
 ## [v0.302.1] - 2025-10-31